cmd/containerd: split package for cli.App
