+++ cmd/containerd/command/config_linux.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	"github.com/containerd/containerd/defaults"
