+++ cmd/containerd/command/config_windows.go
@@ -1,6 +1,4 @@
-// +build !linux,!windows,!solaris
-
-package main
+package command
 
 import (
 	"github.com/containerd/containerd/defaults"
@@ -13,10 +11,6 @@ func defaultConfig() *server.Config {
 		State: defaults.DefaultStateDir,
 		GRPC: server.GRPCConfig{
 			Address: defaults.DefaultAddress,
-		},
-		Debug: server.Debug{
-			Level:   "info",
-			Address: defaults.DefaultDebugAddress,
 		},
 	}
 }
