+++ cmd/containerd/command/publish.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	gocontext "context"
